Add files via upload
can save and log answers; list possible words from a given schema. Different solution methods, including one which chooses the word by ranking according to the smallest average size of the set of words generated by applying it--this doesn't seem to work very well, though... The best approach still seems to be choosing randomly from a slice of the highest-frequency ranked words...With the 3rd word set (smallest), consisting of 2750 or so five letter words, this seem to average around 4.2 attempts.  
I've also made it possible to revise some of the parameters of the searches, including the function used for mode 4, which now defaults to square root...